Refactor: external internal 분리
