+++ payment-api/src/main/java/org/ecommerce/paymentapi/external/service/BeanPayService.java
@@ -1,5 +1,5 @@
 
-package org.ecommerce.paymentapi.service;
+package org.ecommerce.paymentapi.external.service;
 
 import java.util.UUID;
